// moviedb-app-client/src/app/model/Movie.ts
export interface Movie {
    id: number;
    backdropPath: string;
    budget: number;
    title: string;
    overview: string;
    posterPath: string;
    runtime: number;
    releaseDate: string;
    status: string;
    voteAverage: number;
    homepage: string;
    isFavorite: boolean;
    genres: [];
}